명령어 태그별로 파일 분할.
