+++ chip8/CMakeLists.txt
@@ -8,7 +8,15 @@ set( SRC_PATH ${CHIP8_ROOT}/src )
 set(chip8_src_files
         ${SRC_PATH}/typedef.h
         ${SRC_PATH}/hardware/Chip8.cpp
-        ${SRC_PATH}/hardware/Chip8.h )
+        ${SRC_PATH}/hardware/Chip8.h
+
+        # OPCODE 명령어들.
+        ${SRC_PATH}/hardware/opcodes/opcodes_if.cpp
+        ${SRC_PATH}/hardware/opcodes/opcodes_code.cpp
+        ${SRC_PATH}/hardware/opcodes/opcodes_load.cpp
+        ${SRC_PATH}/hardware/opcodes/opcodes_math.cpp
+        ${SRC_PATH}/hardware/opcodes/opcodes_bit.cpp
+        ${SRC_PATH}/hardware/opcodes/opcodes_keyborad.cpp)
 
 add_subdirectory(test)
 add_executable(chip8 src/main.cpp ${chip8_src_files})
