+++ chip8/src/hardware/Chip8.cpp
@@ -100,7 +100,7 @@ void Chip8::nextStep()
 		case 0x0000: // 기타 명령어.
 		{
 			// 명령어 SET 중에 0x0N00. 즉, N 부분을 명령으로 사용하는 코드는 없다. 그래서 마지막만 찾으면 됨.
-			//http://devernay.free.fr/hacks/chip8/C8TECH10.HTM 여기서 opcode 참고.
+			//http://devernay.free.fr/hacks/chip8/C8TECH10.HTM 여기서 opcodes 참고.
 			switch (DecodeOpCodeForth(opCode)) // opCode & 0x000F
 			{
 				case 0x0000:
@@ -120,66 +120,4 @@ void Chip8::nextStep()
 Chip8::~Chip8()
 {
 
-}
-
-
-// [code]
-//JP(jump) addr ( 점프 addr. ) addr = NNN.
-void Chip8::opCode1NNN(WORD opCode)
-{
-	mProgramCounter = opCode & 0x0FFF; // 앞 자리는 해독이 끝났으니. 실제 명령어가 점프할 데이터인 뒷 바이트를 가져온다.
-}
-
-
-// [code]
-//call addr. ( 호출 addr ). addr = NNN. stack에 현재 주소 push.
-void Chip8::opCode2NNN(WORD opCode)
-{
-	mStack.push_back( mProgramCounter ); //stack에 현재 주소 push.
-	mProgramCounter = opCode & 0x0FFF;
-}
-
-// [if]
-//SE (Skip if Equal) Vx, Byte( 같으면_건너띄기 Vx Byte ).Vx =  X. Byte = KK.
-void Chip8::opCode3XKK(WORD opCode)
-{
-    WORD index = (opCode & 0x0F00) >> 8;
-    if(mRegisters[index] == (opCode & 0x00FF))
-    {
-        mProgramCounter+=2;
-    }
-}
-
-// [if]
-//SNE (Skip if not Equal) Vx, Byte (같지_않으면_건너띄기 Vx Byte )  Vx = X. Byte = KK,
-void Chip8::opCode4XKK(WORD opCode)
-{
-    WORD index = (opCode & 0x0F00) >> 8;
-    if(mRegisters[index] != (opCode & 0x00FF))
-    {
-        mProgramCounter += 2;
-    }
-}
-
-// [if]
-//SE (Skip if Equal ) Vx, Vy ( 같으면_건너띄기  Vx Vy ). Vx =  X, Vy = Y
-void Chip8::opCode5XY0(WORD opCode)
-{
-    WORD indexVx = (opCode & 0x0F00) >> 8;
-    WORD indexVy = (opCode & 0x00F0) >> 4;
-    if(mRegisters[indexVx] == mRegisters[indexVy])
-    {
-        mProgramCounter += 2;
-    }
-}
-
-// [load]
-//LD (LoaD) Vx, Byte . ( 로드 Vx Byte ) Vx = X, Byte = KK
-//LD는 Load의 약자다 =ㅁ=.
-void Chip8::opCode6XKK(WORD opCode)
-{
-	BYTE register_index = opCode & 0x0F00;
-	BYTE load_byte =  opCode & 0x00FF;
-
-	mRegisters[register_index] = load_byte;
 }
