+++ chip8/src/hardware/opcodes/opcodes_math.cpp
@@ -1,0 +1,5 @@
+//
+// Created by ffdd270 on 2020-05-20.
+//
+
+#include "../Chip8.h"
